Condensed category commands into one command
